removed all handlers from routes
